+++ swing/framework-ui-test/src/main/java/is/codion/swing/framework/ui/test/EntityEditPanelTestUnit.java
@@ -48,7 +48,7 @@ public class EntityEditPanelTestUnit {
 	 * Initializes the edit panel.
 	 * @throws Exception in case of an exception
 	 */
-	protected final void testInitialize() throws Exception {
+	protected final void testInitialize() {
 		editPanel.apply(editModel).initialize();
 	}